+++ backend/src/main/kotlin/no/nav/historisk/superhelt/ansatt/NavAnsatt.kt
@@ -1,6 +1,6 @@
 package no.nav.historisk.superhelt.ansatt
 
-import no.nav.common.consts.EksternFellesKodeverkTema
+import no.nav.common.consts.FellesKodeverkTema
 import no.nav.common.types.NavIdent
 import no.nav.entraproxy.Enhet
 import no.nav.historisk.superhelt.infrastruktur.authentication.Role
@@ -10,6 +10,6 @@ data class NavAnsatt(
     val ident: NavIdent,
     val roles: List<Role>,
     val enheter: List<Enhet>,
-    val tema: List<EksternFellesKodeverkTema>
+    val tema: List<FellesKodeverkTema>
 
 )
