+++ libs/dokarkiv/src/main/kotlin/no/nav/saf/graphql/DokumentoversiktGraphqlTypes.kt
@@ -1,11 +1,11 @@
 package no.nav.saf.graphql
 
-import no.nav.common.consts.EksternFellesKodeverkTema
+import no.nav.common.consts.FellesKodeverkTema
 
 data class DokumentoversiktFagsakVariables(
     val fagsakId: String,
     val fagsaksystem: String,
-    val tema: List<EksternFellesKodeverkTema>,
+    val tema: List<FellesKodeverkTema>,
     val foerste: Int = 10
 )
