+++ libs/dokarkiv/src/test/kotlin/no/nav/dokarkiv/DokarkivData.kt
@@ -1,6 +1,6 @@
 package no.nav.dokarkiv
 
-import no.nav.common.consts.EksternFellesKodeverkTema
+import no.nav.common.consts.FellesKodeverkTema
 import no.nav.common.types.Enhetsnummer
 import no.nav.common.types.Saksnummer
 import java.io.File
@@ -9,7 +9,7 @@ fun getMockJournalpostRequest() =
     JournalpostRequest(
         tittel = "Dummy Journalpost Title2",
         journalpostType = JournalpostType.UTGAAENDE,
-        tema = EksternFellesKodeverkTema.HEL,
+        tema = FellesKodeverkTema.HEL,
         avsenderMottaker =
             AvsenderMottaker(
                 id = "12345678901",
