+++ mocks/mock-server/src/main/kotlin/no/nav/historisk/mock/oppgave/OppgaveTestData.kt
@@ -1,7 +1,7 @@
 package no.nav.historisk.mock.oppgave
 
 import net.datafaker.Faker
-import no.nav.common.consts.EksternFellesKodeverkTema
+import no.nav.common.consts.FellesKodeverkTema
 import no.nav.common.types.AktorId
 import no.nav.common.types.EksternJournalpostId
 import no.nav.common.types.EksternOppgaveId
@@ -61,7 +61,7 @@ fun generateOppgave(fnr:String?= null, tilordnetRessurs: NavIdent?= null): Oppga
         opprettetAv = faker.numerify("Z######"),
         opprettetTidspunkt = faker.timeAndDate().past().atOffset(ZoneOffset.UTC),
         fristFerdigstillelse = LocalDate.ofInstant(faker.timeAndDate().future(), ZoneOffset.UTC),
-        tema = EksternFellesKodeverkTema.HEL.kode,
+        tema = FellesKodeverkTema.HEL.kode,
         behandlingstema = behandlingstema.kode,
         behandlingstype = behandlingstype?.kode,
         oppgavetype = OppgaveType.JFR.oppgavetype,
